### 2019-01-18 725e69f
removed figure from paper
