### 2019-01-21 #1184 @lheagy
:wave: Hi @fil, @Zac-HD, would you be willing to review this submission for JOSS?
