### 2019-01-21 #1184 @Fil
Sure!

-- Fil

On Mon, Jan 21, 2019 at 6:36 PM Lindsey Heagy <notifications@github.com>
wrote:

> 👋 Hi @Fil <https://github.com/Fil>, @Zac-HD <https://github.com/Zac-HD>,
> would you be willing to review this submission for JOSS?
>
