### 2019-01-27 #1207 @whedon
[REVIEW]: cartoee: Publication quality maps using Earth Engine
**Submitting author:** @kmarkert (<a href="http://orcid.org/0000-0002-7557-0425">Kel Markert</a>)
**Repository:** <a href="https://github.com/kmarkert/cartoee" target ="_blank">https://github.com/kmarkert/cartoee</a>
**Version:** 0.0.5
