### 2019-01-28 #1207 @Fil
I don't believe that there is COI in this case, but just to make sure, let me state that I have been working on a similar tool, written in Javascript and WebGL, on and off for the last few years. You can see an (older) instance of this tool at [URL](https://www.davidrumsey.com/luna/servlet/detail/RUMSEY~8~1~308989~90078901:Urbano-Monte-World-Map-Reprojection?sort=Pub_List_No_InitialSort%2CPub_Date%2CPub_List_No%2CSeries_No&qvq=q:visionscarto.net;sort:Pub_List_No_InitialSort%2CPub_Date%2CPub_List_No%2CSeries_No;lc:RUMSEY~8~1&mi=0&trs=1).

@lheagy @KMarkert feel free to revoke me if this sounds (even slightly) annoying. For my part I see this rather as a good way to learn and maybe cooperate in the future.
