### 2019-01-28 KMarkert/cartoee#1 @Fil
Review for JOSS
Ref. https://github.com/openjournals/joss-reviews/issues/1207

- [x] version number mismatch: you have published release `0.03` (on github) where it should be `0.0.3` (like on pip)

- [x] installation / list of dependencies: I've tried installing with jupyter lab/python3 ~and I get an error from ee indicating that it's probably only compatible with python2. Is this correct, and can you clarify this situation? Please mention that ee is a dependency too (even if that's pretty obvious).~

(That's all I can say for now. I'll be back later if I manage to install it.)
