### 2019-01-28 KMarkert/cartoee#1 @KMarkert
Thanks for the review @Fil!

version number mismatch: I believe that all of the versions between github, pip, and zenodo are version 0.03. Can you please clarify where the versions are mismatched and I will fix the mismatch?

installation / list of dependencies: the ee package should work with python3 (that is what I have been using). The fact that ee is a dependency is mentioned in the documentation [here](https://cartoee.readthedocs.io/en/latest/introduction.html#dependencies) and can be installed using `pip install earthengine-api`. Can you please clarify what the error message was that you received for the installation and I may help.
