### 2019-01-28 KMarkert/cartoee#1 @Fil
The mismatch is the presence of a dot. On github the release is 0.03 and in pip I got cartoee (0.0.3)

`pip install cartoee` did not install all the necessary packages for some reason

Maybe I made a mistake when I thought I should `pip install ee`
now I've tried  `pip install cartoee earthengine-api` instead, and I receive

> ModuleNotFoundError: No module named 'oauth2client'
then same error with cartopy, then

ValueError: numpy.ufunc size changed, may indicate binary incompatibility. Expected 216 from C header, got 192 from PyObject

I'm not necessarily giving up, but not at the point where I can test things on my computer.

Have just tried taking a different route with a colaboratory notebook, and I get the same error again, as if the dependencies were not recognized

![capture d ecran 2019-01-28 a 21 00 19](https://user-images.githubusercontent.com/7001/51862876-e09fed80-233f-11e9-8618-dc9eb94e41cd.png)
