+++ README.md
@@ -1,5 +1,5 @@
 # cartoee
-[![PyPI version](https://badge.fury.io/py/cartoee.svg)](https://badge.fury.io/py/cartoee) [![Build Status](https://travis-ci.com/KMarkert/cartoee.svg?branch=master)](https://travis-ci.com/KMarkert/cartoee) [![Documentation Status](https://readthedocs.org/projects/cartoee/badge/?version=latest)](https://cartoee.readthedocs.io/en/latest/?badge=latest) [![DOI](https://zenodo.org/badge/163705692.svg)](https://zenodo.org/badge/latestdoi/163705692) [![License: GPL v3](https://img.shields.io/badge/License-GPLv3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0)
+[![PyPI version](https://badge.fury.io/py/cartoee.svg)](https://badge.fury.io/py/cartoee) [![Build Status](https://travis-ci.com/KMarkert/cartoee.svg?branch=master)](https://travis-ci.com/KMarkert/cartoee) [![Documentation Status](https://readthedocs.org/projects/cartoee/badge/?version=latest)](https://cartoee.readthedocs.io/en/latest/?badge=latest) [![DOI](https://zenodo.org/badge/163705692.svg)](https://zenodo.org/badge/latestdoi/163705692) [![License: GPL v3](https://img.shields.io/badge/License-GPLv3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0) [![status](http://joss.theoj.org/papers/2eea2105f42bbab1e49cf26d935837c8/status.svg)](http://joss.theoj.org/papers/2eea2105f42bbab1e49cf26d935837c8)
 
 Publication quality maps using [Earth Engine](https://earthengine.google.com/) and [Cartopy](https://scitools.org.uk/cartopy/docs/latest/)!
 ![alt-text](./docs/_static/intro_fig.png)
