### 2019-01-28 982fa6d
adding install_requires to setup.py
