+++ setup.py
@@ -5,7 +5,7 @@ with open("README.md", "r") as fh:
     long_description = fh.read()
 
 setup(name='cartoee',
-      version='0.0.3',
+      version='0.0.4',
       description='Publication quality maps using Earth Engine and Cartopy!',
       long_description=long_description,
       long_description_content_type="text/markdown",
@@ -13,7 +13,16 @@ setup(name='cartoee',
       packages=setuptools.find_packages(),
       author='Kel Markert',
       author_email='kel.markert@gmail.com',
-      license='GNU GPL',
+      license='GNU GPL v3',
       zip_safe=False,
       include_package_data=True,
+      install_requires=[
+          'matplotlib',
+          'Cython',
+          'pyproj'
+          'cartopy',
+          'oauth2client',
+          'google-api-python-client',
+          'earthengine-api',
+      ],
 )
