### 2019-01-28 49da39a
typo fix for setup.py
