+++ setup.py
@@ -20,7 +20,7 @@ setup(name='cartoee',
           'matplotlib',
           'Cython',
           'pyproj',
-          'cartopy',
+          'cartopy==0.16.0',
           'oauth2client',
           'google-api-python-client',
           'earthengine-api',
