### 2019-01-28 1a0c9aa
adding geos library as install requirement
