### 2019-01-28 03ddc9f
adding custom install script
