### 2019-01-28 5ea71a0
making setup.py python2 and python3 compatable
