### 2019-01-28 aaf4bca
fixing syntax errors
