+++ setup.py
@@ -8,11 +8,7 @@ from setuptools.command.install import install
 class eeAuthCommand(install):
     """Customized setuptools install command - prints a friendly greeting."""
     def run(self):
-        print("Installing dependencies...")
-        subprocess("pip install Cython pyproj cartopy oauth2client google-api-python-client earthengine-api")
-
         print("Authorizing the Earth Engine API on your environment...")
-
         import ee
         from ee.ee_exception import EEException
 
@@ -22,11 +18,11 @@ class eeAuthCommand(install):
 
         # if it doesn't work, then authorize an account for Earth Engine to connect to...
         except EEException:
-          subprocess("earthengine authenticate --quiet")
+          subprocess.call(["earthengine","authenticate","--quiet")
 
           authCode = input("Authorization code:")
 
-          subprocess("earthengine authenticate --authorization-code=$authCode")
+          subprocess.call(["earthengine","authenticate","--authorization-code=$authCode"])
 
           # ...then initialize session
           ee.Initialize()
