+++ setup.py
@@ -18,7 +18,7 @@ class eeAuthCommand(install):
 
         # if it doesn't work, then authorize an account for Earth Engine to connect to...
         except EEException:
-          subprocess.call(["earthengine","authenticate","--quiet")
+          subprocess.call(["earthengine","authenticate","--quiet"])
 
           authCode = input("Authorization code:")
 
