### 2019-01-29 ba514df
updating installation required packages
