### 2019-01-29 061b426
creating installation test scripts
