### 2019-01-29 ca54dd2
fixing subprocess call in installation test
