### 2019-01-29 713b9f8
fixing warning printing and error handling
