+++ cartoee/tests/installation_test.py
@@ -5,23 +5,24 @@ import subprocess
 def main():
     try:
         import matplotlib.pyplot as plt
-        print('Successfully imported matplotlib')
+        print('Successfully imported matplotlib\n')
 
     except ImportError:
-        warnings.warn("Could not import matplotlib...")
+        warnings.warn("Could not import matplotlib...\n")
 
     try:
         import ee
         from ee.ee_exception import EEException
-        print('Successfully imported ee')
+        print('Successfully imported ee\n')
 
         # try to initialize Earth Engine session
         try:
             ee.Initialize()
-            print('Successfully initialized and ee session')
+            print('Successfully initialized and ee session\n')
 
         # if it doesn't work, then authorize an account for Earth Engine to connect to...
         except EEException:
+            print('Trying to authorize your ee account...\n')
             subprocess.call('earthengine authenticate --quiet',shell=True)
 
             authCode = input("Authorization code:")
@@ -29,27 +30,31 @@ def main():
             subprocess.call('earthengine authenticate --authorization-code={}'.format(authCode),
                             shell=True)
 
-            # ...then initialize session
-            ee.Initialize()
-            print('Successfully initialized and ee session')
+            try:
+                # ...then initialize session
+                ee.Initialize()
+                print('Successfully initialized and ee session\n')
 
-        finally:
-            warnings.warn("Could not import initialize ee session...")
+            # if it doesn't work, then throw warning...
+            except EEException:
+                warnings.warn("Could not import initialize ee session...\n")
 
     except ImportError:
-        warnings.warn("Could not import ee...")
+        warnings.warn("Could not import ee...\n")
 
     try:
         import cartopy
         import cartopy.crs as ccrs
-        print('Successfully imported cartopy')
+        print('Successfully imported cartopy\n')
 
     except:
-        warnings.warn("Could not import cartopy...")
+        warnings.warn("Could not import cartopy...\n")
 
     try:
         import cartoee as cee
-        print('Successfully imported cartoee')
+        print('Successfully imported cartoee\n')
 
     except:
-        warnings.warn("Could not import cartoee...")
+        warnings.warn("Could not import cartoee...\n")
+
+    print('Installation testing done.')
