### 2019-01-29 e53760d
added summary of installation test at end
