### 2019-01-29 f6ed024
adding plotting test script
