### 2019-01-29 3079db8
adding additional plotting tests
