+++ cartoee/tests/plotting_test.py
@@ -13,6 +13,7 @@ def mapTest(img,box,vis):
     ax = cee.getMap(img,region=box,visParams=vis,
                     cmap='gist_earth')
     plt.show()
+    plt.close()
 
 
 def layerTest(img,box,vis):
@@ -22,6 +23,7 @@ def layerTest(img,box,vis):
     ax = cee.addLayer(img,ax=ax,region=box,visParams=vis,
                     cmap='gist_earth')
     plt.show()
+    plt.close()
 
 
 def colorbarTest(img,box,vis):
@@ -31,6 +33,7 @@ def colorbarTest(img,box,vis):
     cb = cee.addColorbar(ax,loc='right',cmap='gist_earth',visParams=vis)
 
     plt.show()
+    plt.close()
 
 
 def projectionTest(img,box,vis):
@@ -44,6 +47,7 @@ def projectionTest(img,box,vis):
     cb = cee.addColorbar(ax,loc='bottom',cmap='gist_earth',visParams=vis)
 
     plt.show()
+    plt.close()
 
 
 def main():
@@ -51,9 +55,44 @@ def main():
     bbox = [-180,-60,180,90]
     srtm = ee.Image("CGIAR/SRTM90_V4")
 
-    print("Testing basin plotting functionality...")
+    print("Testing getMap functionality...")
     try:
         mapTest(srtm,bbox,visualization)
-        print('Plotting successful')
+        print('getMap test successful')
+        t1 = 'successful'
     except Exception as e:
-        warnings.warn("Plotting test failed...")
+        warnings.warn("getMap test failed...")
+        t1 = 'failed'
+
+    print("Testing addLayer functionality...")
+    try:
+        layerTest(srtm,bbox,visualization)
+        print('addLayer test successful')
+        t2 = 'successful'
+    except Exception as e:
+        warnings.warn("addLayer test failed...")
+        t2 = 'failed'
+
+    print("Testing colorbar functionality...")
+    try:
+        colorbarTest(srtm,bbox,visualization)
+        print('colorbar test successful')
+        t3 = 'successful'
+    except Exception as e:
+        warnings.warn("colorbar test failed...")
+        t3 = 'failed'
+
+    print("Testing projection functionality...")
+    try:
+        projectionTest(srtm,bbox,visualization)
+        print('projection test successful')
+        t4 = 'successful'
+    except Exception as e:
+        warnings.warn("colorbar test failed...")
+        t4 = 'failed'
+
+    print('Plotting testing done.\n '
+          'getMap:      {0} \n '
+          'addLayer:    {1} \n '
+          'colorbar:    {2} \n '
+          'projections: {3} \n '.format(t1,t2,t3,t4))
