### 2019-01-29 4e7f11d
removing new figure within each test
