### 2019-01-29 a8d3ce5
fixing minor syntax error
