### 2019-01-29 ae2d9e5
adding in additional installation instructions
