### 2019-01-29 KMarkert/cartoee#1 @KMarkert
Thank you for pointing out that detail. Please confirm that I have fixed all of the mismatched versions (should be 0.0.5 now).

As for the installation, I have added into the setup.py the `install_requires` argument so that all of the Python dependencies will be installed as well. However, the cartopy package will need GEOS and Proj4 libraries installed prior to running `pip install cartoee`. If those libraries are not installed the cartoee install will fail; I have noted this in the [documentation](https://cartoee.readthedocs.io/en/latest/introduction.html#dependencies). The "ValueError: numpy.ufunc size changed" error was also due to dependency's latest version, I fixed the required version of cartopy to be 0.16.0 to circumvent that error for now.

An example Colab notebook for installation and testing the installation can be found at this link: https://colab.research.google.com/drive/1pvDOKC7QLLeOGGLBlFgZwcp6FNde1DhI

Thank you for bearing with me on the installation part.
