### 2019-01-30 KMarkert/cartoee#1 @Fil
Excellent! No problem, it's my job as a JOSS reviewer to make sure things install smoothly.
