### 2019-01-30 #1207 @lheagy
Many thanks @Fil for the speedy review!!

@KMarkert, could you please make a new release and ensure that the title of the zenodo archive is the same as your paper "cartoee: Publication quality maps using Earth Engine" and post the doi here? From there, we can proceed with publication 🎉
