### 2019-01-30 #1207 @KMarkert
@Fil Thank you very much for the reviews and helping make the package of higher quality.

@lheagy I have edited the name of the zenodo archive to reflect the name of the paper. Can you please confirm that I did this correctly? Here is the doi: https://dx.doi.org/10.5281/zenodo.2552811
