### 2019-01-30 #1207 @danielskatz
👋 @arfon - I don't see the paper itself, which should be at https://www.theoj.org/joss-papers/joss.01207/10.21105.joss.01207.pdf
Can you see what might have gone wrong?
