### 2019-01-30 #1207 @arfon
> Can you see what might have gone wrong?

The paper seems to be there now. Sometimes GitHub pages (where the PDFs are served from) takes longer than a few seconds to update.
